backend/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
python-socketio==5.10.0
eventlet==0.33.3
PyJWT==2.8.0
python-dotenv==1.0.0
pynput==1.7.6
comtypes==1.4.0
Pillow==10.1.0
requests==2.31.0
psutil==5.9.6
cryptography==41.0.7
pytest==7.4.3
pytest-flask==1.3.0

